[ISSUE #3815]DelayRetryable and RetryContext can be refactored. (#3819)
* Refacotr DelayRetryable and RetryContext.

* Modify as review.
