+++ eventmesh-runtime/src/main/java/org/apache/eventmesh/runtime/core/protocol/grpc/push/AbstractPushRequest.java
@@ -28,9 +28,9 @@ import org.apache.eventmesh.protocol.api.ProtocolPluginFactory;
 import org.apache.eventmesh.runtime.boot.EventMeshGrpcServer;
 import org.apache.eventmesh.runtime.configuration.EventMeshGrpcConfiguration;
 import org.apache.eventmesh.runtime.constants.EventMeshConstants;
+import org.apache.eventmesh.runtime.core.protocol.RetryContext;
 import org.apache.eventmesh.runtime.core.protocol.grpc.consumer.EventMeshConsumer;
 import org.apache.eventmesh.runtime.core.protocol.grpc.retry.GrpcRetryer;
-import org.apache.eventmesh.runtime.core.protocol.grpc.retry.RetryContext;
 
 import java.util.Collections;
 import java.util.Map;
@@ -101,9 +101,8 @@ public abstract class AbstractPushRequest extends RetryContext {
     }
 
     @Override
-    public boolean retry() {
+    public void retry() {
         tryPushRequest();
-        return true;
     }
 
     protected void delayRetry() {
